Finished implementing comments on news

+++ controllers/newsCommentsController.js
@@ -31,9 +31,48 @@ module.exports = {
         }
     },
 
+    async commentsEditPatch(req, res, next) {
+        const { comments_id } = req.params;
+        const { comments_content } = req.body;
+
+        const comment = await newsCommentsModel.getNewsCommentById(comments_id);
+
+        if(!comment) {
+            const error = new Error("Comment not found.");
+            error.status = 404;
+
+            return next(error);
+        }
+
+        if(comment.comments_author != res.locals.sessionUser.userId) {
+            res.status(401);
+            return res.redirect("/");
+        }
+
+        await newsCommentsModel.updateCommentById(comments_content, comments_id);
+
+        res.status(200);
+
+        return res.json({ success: true, comments_content: comments_content });
+    },
+
     async commentsDelete(req, res, next) {
         const { comments_id } = req.params;
         const { news_slug } = req.body;
+
+        const comment = await newsCommentsModel.getNewsCommentById(comments_id);
+
+        if(!comment) {
+            const error = new Error("Not found");
+            error.status = 404;
+
+            return next(error);
+        }
+
+        if(comment.comments_author != res.locals.sessionUser.userId) {
+            res.status(401);
+            return res.redirect("/");
+        }
 
         try {
             await newsCommentsModel.deleteNewsCommentById(comments_id);
